refactor(service): 更新动态统计信息方法入参类型修改
-将 updatePostStatistics 方法的参数类型从 Long 修改为 Integer
-此修改统一了方法参数类型，与其他方法保持一致

+++ backend/src/main/java/cn/hmck/service/impl/PostServiceImpl.java
@@ -88,7 +88,7 @@ public class PostServiceImpl extends ServiceImpl<PostMapper, Post> implements Po
 
     // 更新动态统计信息
     @Override
-    public void updatePostStatistics(Long postId) {
+    public void updatePostStatistics(Integer postId) {
         // 统计点赞数和评论数
         int likeCount = commentMapper.countCommentsByPostIdAndType(postId, 2);  // 获取点赞数，type = 2
         int commentCount = commentMapper.countCommentsByPostIdAndType(postId, 1);  // 获取评论数，type = 1
